added go bin for windows, linux and mac
