--- kafka/go.mod
@@ -1,5 +1,0 @@
-module github.com/maximilianokaizen/yape-fork
-
-go 1.18
-
-require github.com/confluentinc/confluent-kafka-go v1.9.2 // indirect
